Game board 4
I made the grid so when the user clicks on the top row, the squares are
red, and elsewhere they are yellow.  I need to get them to stay these
colors and test when the player has all the red ones hit.  I also need
to go to a computer grid, guessing where the user has placed ships.